updated type for amount

+++ migrations/20220328044357-create-restaurant.js
@@ -13,7 +13,7 @@ module.exports = {
         allowNull: false
       },
       cashBalance: {
-        type: Sequelize.DECIMAL,
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       createdAt: {
