+++ migrations/20220328044618-create-user.js
@@ -12,7 +12,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       cashBalance: {
-        type: Sequelize.DECIMAL
+        type: Sequelize.FLOAT
       },
       createdAt: {
         allowNull: false,
